Rename bound_args_transform->config
