+++ cyclopts/__init__.py
@@ -15,7 +15,7 @@ __all__ = [
     "UnknownOptionError",
     "UnusedCliTokensError",
     "ValidationError",
-    "bound_args_transforms",
+    "config",
     "convert",
     "default_name_transform",
     "types",
@@ -40,4 +40,4 @@ from cyclopts.parameter import Parameter
 from cyclopts.protocols import Dispatcher
 from cyclopts.utils import default_name_transform
 
-from . import bound_args_transforms, types, validators
+from . import config, types, validators
