+++ tests/config/test_end2end.py
@@ -2,11 +2,11 @@ from textwrap import dedent
 
 import pytest
 
-from cyclopts.bound_args_transforms import Toml
+from cyclopts.config import Toml
 
 
 @pytest.mark.skip(reason="Need to think carefully about where injection is performed.")
-def test_bound_args_transform_end2end(app, tmp_path):
+def test_config_end2end(app, tmp_path):
     config_fn = tmp_path / "config.toml"
 
     config_fn.write_text(
@@ -23,7 +23,7 @@ def test_bound_args_transform_end2end(app, tmp_path):
         )
     )
 
-    app.bound_args_transform = Toml(config_fn, root_keys=["tool", "cyclopts"])
+    app.config = Toml(config_fn, root_keys=["tool", "cyclopts"])
 
     @app.default
     def default(key1, key2):
